feat(#699): add AsmClass class
